Chapter 3: Add shared L1 Cache
For L1 cache to be shared between SMT threads, a register array must be
used. This, however, is not straightforward if every node in the CPU map
refers to a separate CPU node. Therefore, it is suggested to create a
separate CPU node for every SMT thread. The L1 cache can be shared if an
extra node represents it.

Add an extra layer for shared L1 cache in the device tree among multiple
SMTs.

Implementation associated with these changes has already been posted at
[1].

Link: https://lore.kernel.org/all/20250129164855.676-1-alireza.sanaee@huawei.com # [1]
Signed-off-by: Alireza Sanaee <alireza.sanaee@huawei.com>